pyproject.toml
[tool.poetry]
name = "metricflow"
version = "0.91.8"
description = ""
authors = ["Transform <hello@transformdata.io>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"

croniter = "^1.3.4"
pycron = "^3.0.0"
pydantic = "^1.9.0"
fuzzywuzzy = "^0.18.0"
jsonschema = "3.2.0"
SQLAlchemy = "^1.4.27"
snowflake-sqlalchemy = "1.2.3"
sqlalchemy-redshift = "0.8.1"
numpy = "1.22.0"
pandas = "1.2.4"
Jinja2 = "2.11.3"
PyYAML = "^5.4.1"
snowflake-connector-python = "^2.7.6"
moz-sql-parser = "^4.40.21126"
tabulate = "0.8.9"
more-itertools = "8.10.0"
graphviz = "0.18.2"
sqlalchemy2-stubs = "^0.0.2-alpha.21"
sqlalchemy-bigquery = "^1.4.3"
python-dateutil = "2.8.2"
requests = "^2.27.1"
MarkupSafe = "2.0.1"
psycopg2 = "^2.9.3"
google-cloud-bigquery = "2.34.2"
halo = "^0.0.31"
update-checker = "^0.18.0"
ciso8601 = "^2.2.0"
click = "^7.1.2"
"ruamel.yaml" = "^0.17.21"
python-Levenshtein = "^0.12.2"
rudder-sdk-python = "^1.0.3"

[tool.poetry.dev-dependencies]
pytest-mock = "^3.7.0"
pytest = "^7.1.1"
pre-commit = "^2.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mf = 'metricflow.cli.main:cli'

[tool.black]
line-length = 120
target-version = ['py36', 'py37', 'py38']
force-exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''
